Revert Pure node, add namedFn
Pure is rather complex and still does not get us to kleisli. By adding a
namedFn node where the user gives a unique id for each function, we can
simplify a lot. It is obviously risky, but we already assume Monad
coherence, so taking functions like pure or flatMap to be stable, is not
really increasing risk.

We need some kind of clean/invalidation to deal with cases where there
are real changes to the monad that has already cached output.